pyproject.toml
[project]
name = "lbrxvoicepro"
version = "1.0.0"
description = "Universal Voice AI Pipeline - Dataset collection in MOSHI/MIMI format, MLX-optimized ASR/TTS, and conversational AI with universal RAG"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [{name = "LIBRAXIS Team", email = "team@libraxis.com"}]
dependencies = [
    "chromadb>=0.4.0",
    "fastapi>=0.115.0",
    "httpx>=0.28.0",
    "mlx>=0.26.0",
    "mlx-whisper>=0.4.0",
    "numpy>=1.26.0",
    "pydantic>=2.0.0",
    "python-multipart>=0.0.18",
    "rich>=13.0.0",
    "ruff>=0.12.0",
    "scipy>=1.12.0",
    "sentence-transformers>=3.0.0",
    "silero-vad>=5.0",
    "sounddevice>=0.5.0",
    "soundfile>=0.12.0",
    "torch>=2.0.0",
    "uvicorn>=0.32.0",
    "websockets>=14.0",
]

[dependency-groups]
dev = [
    "mypy>=1.16.0",
    "pytest>=8.4.0",
    "pytest-asyncio>=1.0.0",
    "pytest-cov>=6.1.1",
    "ruff>=0.11.12",
]
